mongoose connection moved to config/connectDB, added useCreateIndex, applied those updates to server and seed files and changed mongoose ver to 5.11.15 to eliminate circular dep error
