--- config/config.example
@@ -1,4 +1,0 @@
-PORT = 3000
-MONGO_URI = xxxxx
-GOOGLE_CLIENT_ID = xxxxx
-GOOGLE_CLIENT_SECRET = xxxxx
