--- config/db.js
@@ -1,19 +1,0 @@
-// const mongoose = require("mongoose");
-
-// const connectDB = async () => {
-//   try {
-//     const conn = await mongoose.connect(process.env.MONGO_URI, {
-//       useNewUrlParser: true,
-//       useUnifiedTopology: true,
-//       useFindAndModify: false,
-//     })
-  
-//       console.log(`MongoDB Connected: ${conn.connection.host}`)
-//   } catch (err) {
-//       console.error(err)
-//       process.exit(1)
-//   }
-// }
-
-// module.exports = connectDB;
-
